refactor(cli): split render pipeline modules
